+++ dds/src/xtypes/deserializer.rs
@@ -1,8 +1,8 @@
 use crate::xtypes::{
-    data_storage::F128, dynamic_type::{
+    dynamic_type::{
         DynamicData, DynamicDataFactory, DynamicType, DynamicTypeMember, ExtensibilityKind,
         TypeKind,
-    }, error::{XTypesError, XTypesResult}, read_write::Read
+    }, error::{XTypesError, XTypesResult}, f128::F128, read_write::Read
 };
 use alloc::{string::String, vec::Vec};
 use tracing::debug;
